chore(angular): move type to .storybook
